Use batches not full data in batch irls

+++ projects/project1/group3/Narozniak_Czerwonski_Rucinski/src/irls_optimizer.py
@@ -4,7 +4,7 @@ import numpy as np
 class IRLS:
     def update(self, B,  X, y):
         # Add 1 as the first column to X
-        X_ = np.concatenate([np.ones(shape=(X.shape[0], 1)), X], axis=1)
+        X_ = np.concatenate([np.ones(shape=(X.shape[0], 1)).astype(float), X], axis=1)
         z = (X_ @ B).astype(float)
         p = 1 / (1 + np.exp(-z))
         W = np.diagflat(p * (1 - p))
